requirements.txt

click
PyPDF2
python-docx
openai
langchain
google-generativeai
promptl-ai
python-dotenv
